--- src/main/java/roomescape/dto/response/reservation/MyReservationWebResponse.java
@@ -1,35 +1,0 @@
-package roomescape.dto.response.reservation;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.Status;
-
-public record MyReservationWebResponse(
-        Long id,
-        String theme,
-        LocalDate date,
-        @JsonFormat(pattern = "HH:mm")
-        LocalTime time,
-        String status,
-        String paymentKey,
-        BigDecimal totalAmount
-) {
-    public static MyReservationWebResponse of(Reservation reservation, String paymentKey, BigDecimal totalAmount,
-                                              long rank) {
-        return new MyReservationWebResponse(
-                reservation.getId(), reservation.getTheme().getName(), reservation.getDate(),
-                reservation.getTime().getStartAt(), getWaitingOrder(reservation.getStatus(), rank),
-                paymentKey, totalAmount
-        );
-    }
-
-    private static String getWaitingOrder(Status status, long rank) {
-        if (status == Status.WAITING) {
-            return rank + "번째 " + status.getValue();
-        }
-        return status.getValue();
-    }
-}
