+++ src/main/java/roomescape/service/ReservationService.java
@@ -16,7 +16,7 @@ import roomescape.dto.request.reservation.ReservationInformRequest;
 import roomescape.dto.request.reservation.ReservationRequest;
 import roomescape.dto.request.reservation.WaitingRequest;
 import roomescape.dto.response.reservation.CanceledReservationWebResponse;
-import roomescape.dto.response.reservation.MyReservationWebResponse;
+import roomescape.dto.response.reservation.MyReservationResponse;
 import roomescape.dto.response.reservation.ReservationInformResponse;
 import roomescape.dto.response.reservation.ReservationResponse;
 
@@ -97,9 +97,9 @@ public class ReservationService {
     }
 
     @Transactional(readOnly = true)
-    public List<MyReservationWebResponse> findMyReservations(Long memberId) {
+    public List<MyReservationResponse> findMyReservations(Long memberId) {
         return reservationRepository.findMyReservation(memberId).stream()
-                .map(myReservationResponse -> MyReservationWebResponse.of(
+                .map(myReservationResponse -> MyReservationResponse.of(
                         myReservationResponse.reservation(), myReservationResponse.paymentKey(),
                         myReservationResponse.totalAmount(), getWaitingOrder(myReservationResponse.reservation())))
                 .toList();
