+++ src/test/java/roomescape/TestFixtures.java
@@ -10,7 +10,7 @@ import roomescape.dto.request.reservation.AdminReservationRequest;
 import roomescape.dto.response.reservation.AvailableTimeResponse;
 import roomescape.dto.response.member.MemberResponse;
 import roomescape.dto.request.member.MemberSignUpRequest;
-import roomescape.dto.response.reservation.MyReservationWebResponse;
+import roomescape.dto.response.reservation.MyReservationResponse;
 import roomescape.dto.request.reservation.ReservationCriteriaRequest;
 import roomescape.dto.response.reservation.ReservationResponse;
 import roomescape.dto.request.reservation.ReservationTimeRequest;
@@ -197,25 +197,25 @@ public class TestFixtures {
             RESERVATION_RESPONSE_10, RESERVATION_RESPONSE_11, RESERVATION_RESPONSE_12, RESERVATION_RESPONSE_13,
             RESERVATION_RESPONSE_14, RESERVATION_RESPONSE_15
     );
-    public static final MyReservationWebResponse MY_RESERVATION_RESPONSE_1 = new MyReservationWebResponse(
+    public static final MyReservationResponse MY_RESERVATION_RESPONSE_1 = new MyReservationResponse(
             1L, "name1", TODAY.minusDays(5), LocalTime.of(10, 0), "예약", null, null
     );
-    public static final MyReservationWebResponse MY_RESERVATION_RESPONSE_2 = new MyReservationWebResponse(
+    public static final MyReservationResponse MY_RESERVATION_RESPONSE_2 = new MyReservationResponse(
             2L, "name1", TODAY.minusDays(5), LocalTime.of(11, 0), "예약", null, null
     );
-    public static final MyReservationWebResponse MY_RESERVATION_RESPONSE_3 = new MyReservationWebResponse(
+    public static final MyReservationResponse MY_RESERVATION_RESPONSE_3 = new MyReservationResponse(
             3L, "name1", TODAY.minusDays(5), LocalTime.of(12, 0), "예약", null, null
     );
-    public static final MyReservationWebResponse MY_RESERVATION_RESPONSE_4 = new MyReservationWebResponse(
+    public static final MyReservationResponse MY_RESERVATION_RESPONSE_4 = new MyReservationResponse(
             4L, "name1", TODAY.minusDays(5), LocalTime.of(13, 0), "예약", null, null
     );
-    public static final MyReservationWebResponse MY_RESERVATION_RESPONSE_5 = new MyReservationWebResponse(
+    public static final MyReservationResponse MY_RESERVATION_RESPONSE_5 = new MyReservationResponse(
             5L, "name2", TODAY.minusDays(5), LocalTime.of(10, 0), "예약", null, null
     );
-    public static final MyReservationWebResponse MY_RESERVATION_RESPONSE_6 = new MyReservationWebResponse(
+    public static final MyReservationResponse MY_RESERVATION_RESPONSE_6 = new MyReservationResponse(
             6L, "name2", TODAY.minusDays(5), LocalTime.of(11, 0), "예약", null, null
     );
-    public static final List<MyReservationWebResponse> MY_RESERVATION_RESPONSES = List.of(
+    public static final List<MyReservationResponse> MY_RESERVATION_RESPONSES = List.of(
             MY_RESERVATION_RESPONSE_1, MY_RESERVATION_RESPONSE_2, MY_RESERVATION_RESPONSE_3,
             MY_RESERVATION_RESPONSE_4, MY_RESERVATION_RESPONSE_5, MY_RESERVATION_RESPONSE_6
     );
